chore: implement ESLint rulset as a package in the workspace
