+++ package.json
@@ -120,8 +120,8 @@
         "@babel/preset-react": "^7.24.7",
         "@babel/preset-typescript": "^7.24.7",
         "@babel/runtime": "^7.23.9",
-        "@eslint/js": "^9.13.0",
         "@suite-common/wallet-types": "workspace:*",
+        "@trezor/eslint": "workspace:*",
         "@types/jest": "29.5.12",
         "@types/node": "20.12.7",
         "@types/prettier": "^3.0.0",
@@ -129,11 +129,6 @@
         "@types/tar": "^6.1.11",
         "babel-jest": "29.7.0",
         "depcheck": "^1.4.7",
-        "eslint": "^9.13.0",
-        "eslint-plugin-import": "^2.31.0",
-        "eslint-plugin-jest": "^28.8.3",
-        "eslint-plugin-react": "^7.37.2",
-        "eslint-plugin-react-hooks": "^5.0.0",
         "globals": "^15.11.0",
         "jest": "29.7.0",
         "jest-environment-jsdom": "29.7.0",
@@ -150,7 +145,6 @@
         "tslib": "^2.6.2",
         "tsx": "^4.16.3",
         "typescript": "5.5.4",
-        "typescript-eslint": "^8.12.2",
         "version-bump-prompt": "^6.1.0"
     },
     "dependenciesMeta": {
