+++ packages/connect-examples/webextension-mv2/eslint.config.mjs
@@ -1,3 +1,4 @@
+// Todo: figure out the better way, if this would be package, we would use "@trezor/eslint": "workspace:*" in package.json
 import parentConfig from '../../../eslint.config.mjs';
 
 export default [
