+++ packages/connect-explorer/package.json
@@ -48,6 +48,7 @@
         "swr": "^2.2.5"
     },
     "devDependencies": {
+        "@trezor/eslint": "workspace:*",
         "@types/redux-logger": "^3.0.11",
         "babel-plugin-styled-components": "^2.1.4",
         "concurrently": "^8.2.2",
