+++ packages/connect-webextension/eslint.config.mjs
@@ -1,7 +1,7 @@
-import parentConfig from '../../eslint.config.mjs';
+import { eslint } from '@trezor/eslint';
 
 export default [
-    ...parentConfig,
+    ...eslint,
     {
         rules: {
             'no-underscore-dangle': 'off', // underscore is used
