+++ packages/connect-webextension/package.json
@@ -46,6 +46,7 @@
     },
     "devDependencies": {
         "@babel/preset-typescript": "^7.24.7",
+        "@trezor/eslint": "workspace:*",
         "@trezor/node-utils": "workspace:*",
         "@trezor/trezor-user-env-link": "workspace:*",
         "@types/chrome": "^0.0.278",
