+++ packages/eslint/src/index.mjs
@@ -1,0 +1,1 @@
+export { default as eslint } from './eslint.config.mjs';
