+++ packages/schema-utils/package.json
@@ -23,6 +23,7 @@
     },
     "devDependencies": {
         "@sinclair/typebox-codegen": "^0.10.4",
+        "@trezor/eslint": "workspace:*",
         "ts-node": "^10.9.2",
         "tsx": "^4.16.3"
     },
