+++ packages/transport-test/package.json
@@ -20,6 +20,7 @@
     },
     "devDependencies": {
         "@jest/types": "^29.6.3",
+        "@trezor/eslint": "workspace:*",
         "@trezor/transport": "workspace:*",
         "@trezor/transport-bridge": "workspace:*",
         "@trezor/trezor-user-env-link": "workspace:^",
