+++ packages/utils/package.json
@@ -34,6 +34,7 @@
         "tslib": "^2.6.2"
     },
     "devDependencies": {
+        "@trezor/eslint": "workspace:*",
         "tsx": "^4.16.3"
     },
     "dependencies": {
