+++ scripts/package.json
@@ -19,6 +19,7 @@
         "yargs": "17.7.2"
     },
     "devDependencies": {
+        "@trezor/eslint": "workspace:*",
         "@types/prettier": "^3.0.0"
     }
 }
